穿装备, 脱装备 可用

+++ DB/src/main/java/com/zhaoxiaodan/mirserver/db/objects/BaseObject.java
@@ -192,7 +192,7 @@ public abstract class BaseObject {
 		return NumUtil.makeLong(NumUtil.makeWord(race.id, (byte) 0), appr);
 	}
 
-	public int getFeatureEx() {
+	public short getFeatureEx() {
 		return 0;
 	}
